api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["benleong0 <benleong0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
Flask = "^3.0.2"
flask-cors = "^4.0.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.356"
ruff = "^0.3.4"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
include_trailing_comma = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["B023"]

per-file-ignores."testing/dicts.py" = ["E501"]
per-file-ignores."testing/example_script.py" = ["E501"]
per-file-ignores."testing/jisho_api_responses.py" = ["E501"]
per-file-ignores."testing/wanikani_api_responses.py" = ["E501"]
